Refactor: move color picker widget update functions to a separate file.
This is a preparation to use these functions in mwseMCMColorPicker.

Also extract some shared variables into constants and uiid files.

Use sDone GMST value for better localisation.
