+++ misc/package/Data Files/MWSE/core/mwse/ui/tes3uiElement/createColorPicker/constants.lua
@@ -1,0 +1,10 @@
+return {
+	PICKER_HEIGHT = 256,
+	PICKER_MAIN_WIDTH = 256,
+	PICKER_VERTICAL_COLUMN_WIDTH = 32,
+	PICKER_PREVIEW_WIDTH = 64,
+	PICKER_PREVIEW_HEIGHT = 32,
+	INDICATOR_TEXTURE = "textures\\menu_map_smark.dds",
+	INDICATOR_COLOR = { 0.5, 0.5, 0.5 },
+	SLIDER_SCALE = 1000,
+}
