+++ misc/package/Data Files/MWSE/core/mwse/ui/tes3uiElement/createColorPicker/init.lua
@@ -1,33 +1,15 @@
 local ffi = require("ffi")
 
 local ColorPicker = require("mwse.ui.tes3uiElement.createColorPicker.ColorPicker")
+local CONSTANTS = require("mwse.ui.tes3uiElement.createColorPicker.constants")
 local format = require("mwse.ui.tes3uiElement.createColorPicker.formatHelpers")
 local oklab = require("mwse.ui.tes3uiElement.createColorPicker.oklab")
 local premultiply = require("mwse.ui.tes3uiElement.createColorPicker.premultiply")
-
-local PICKER_HEIGHT = 256
-local PICKER_MAIN_WIDTH = 256
-local PICKER_VERTICAL_COLUMN_WIDTH = 32
-local PICKER_PREVIEW_WIDTH = 64
-local PICKER_PREVIEW_HEIGHT = 32
-local INDICATOR_TEXTURE = "textures\\menu_map_smark.dds"
-local INDICATOR_COLOR = { 0.5, 0.5, 0.5 }
-local SLIDER_SCALE = 1000
+local UIID = require("mwse.ui.tes3uiElement.createColorPicker.uiid")
+local update = require("mwse.ui.tes3uiElement.createColorPicker.updateHelpers")
 
 -- Defined in oklab\init.lua
 local ffiPixel = ffi.typeof("RGB") --[[@as fun(init: ffiImagePixelInit?): ffiImagePixel]]
-
-
-local UIID = {
-	menu = tes3ui.registerID("testing:Menu"),
-	pickerMenu = tes3ui.registerID("testing:pickerMenu"),
-	indicator = {
-		main = tes3ui.registerID("ColorPicker_main_picker_indicator"),
-		hue = tes3ui.registerID("ColorPicker_hue_picker_indicator"),
-		alpha = tes3ui.registerID("ColorPicker_alpha_picker_indicator"),
-		slider = tes3ui.registerID("ColorPicker_main_picker_slider"),
-	}
-}
 
 -- TODO: these could use localization.
 local strings = {
@@ -44,125 +26,6 @@ local strings = {
 --- @field standardPreview tes3uiElement
 --- @field checkersPreview tes3uiElement
 
----------------------------
---- UI update functions ---
----------------------------
-
---- @param picker ColorPicker
---- @param previews ColorPickerPreviewsTable
---- @param newColor ffiImagePixel
---- @param alpha number
-local function updatePreview(picker, previews, newColor, alpha)
-	previews.standardPreview.color = { newColor.r, newColor.g, newColor.b }
-	previews.standardPreview:updateLayout()
-	picker:updatePreviewImage(newColor, alpha)
-	previews.checkersPreview.texture.pixelData:setPixelsFloat(picker.previewImage:toPixelBufferFloat())
-end
-
---- @alias IndicatorID
----| "main"
----| "hue"
----| "alpha"
----| "slider"
-
---- @param parent tes3uiElement
-local function getIndicators(parent)
-	--- @type table<IndicatorID, tes3uiElement>
-	local indicators = {}
-	for _, UIID in pairs(UIID.indicator) do
-		local indicator = parent:findChild(UIID)
-		-- Not every Color Picker will have alpha indicator.
-		if indicator then
-			local id = indicator:getLuaData("indicatorID")
-			indicators[id] = indicator
-		end
-	end
-	return indicators
-end
-
---- @param parent tes3uiElement
---- @param newColor ffiImagePixel
---- @param alpha number?
-local function updateIndicatorPositions(parent, newColor, alpha)
-	local hsv = oklab.hsvlib_srgb_to_hsv(newColor)
-	local indicators = getIndicators(parent)
-	indicators.main.absolutePosAlignX = hsv.s
-	indicators.main.absolutePosAlignY = 1 - hsv.v
-	indicators.hue.absolutePosAlignY = hsv.h / 360
-	if indicators.alpha then
-		indicators.alpha.absolutePosAlignY = 1 - alpha
-	end
-	-- Update main picker's slider
-	indicators.slider.widget.current = hsv.s * SLIDER_SCALE
-	indicators.hue:getTopLevelMenu():updateLayout()
-end
-
---- @param parent tes3uiElement
---- @param newColor ffiImagePixel|ImagePixel|ImagePixelA
---- @param alpha number
-local function updateValueInput(parent, newColor, alpha)
-	local input = parent:findChild(tes3ui.registerID("ColorPicker_data_row_value_input"))
-	if not input then return end
-
-	-- Make sure we don't get NaNs in color text inputs. We clamp alpha here.
-	if alpha then
-		alpha = math.clamp(alpha, 0.0000001, 1.0)
-	end
-
-	local newText = ""
-	if input:getLuaData("hasAlpha") then
-		newText = format.pixelToHex({
-			r = newColor.r,
-			g = newColor.g,
-			b = newColor.b,
-			a = alpha,
-		})
-	else
-		newText = format.pixelToHex({
-			r = newColor.r,
-			g = newColor.g,
-			b = newColor.b,
-		})
-	end
-	input.text = newText
-end
-
---- Used to update current preview color and the text shown in the value input.
---- Usually used when after a color was picked in the main or alpha pickers.
---- @param picker ColorPicker
---- @param parent tes3uiElement
---- @param newColor ffiImagePixel
---- @param alpha number
---- @param previews ColorPickerPreviewsTable
-local function colorSelected(picker, parent, newColor, alpha, previews)
-	-- Make sure we don't create reference to the pixel picked from the mainImage.
-	-- We construct a new ffiPixel.
-	newColor = ffiPixel({ newColor.r, newColor.g, newColor.b })
-	picker:setColor(newColor, alpha)
-	updatePreview(picker, previews, newColor, alpha)
-	updateValueInput(parent, newColor, alpha)
-end
-
---- Used when a color with different Hue was picked.
---- @param picker ColorPicker
---- @param parent tes3uiElement
---- @param newColor ffiImagePixel|ImagePixel
---- @param alpha number
---- @param previews ColorPickerPreviewsTable
---- @param mainPicker tes3uiElement
-local function hueChanged(picker, parent, newColor, alpha, previews, mainPicker)
-	-- Make sure we don't create reference to the pixel picked from the mainImage.
-	-- We construct a new ffiPixel.
-	newColor = ffiPixel({ newColor.r, newColor.g, newColor.b })
-	colorSelected(picker, parent, newColor, alpha, previews)
-	-- Now, also need to regenerate the image for the main picker since the Hue changed.
-	picker:updateMainImage(newColor)
-	mainPicker.texture.pixelData:setPixelsFloat(picker.mainImage:toPixelBufferFloat())
-end
-
------------------------------
---- UI creation functions ---
------------------------------
 
 --- @param picker ColorPicker
 --- @param parent tes3uiElement
@@ -264,16 +127,16 @@ end
 local function createIndicator(parent, id, absolutePosAlignX, absolutePosAlignY)
 	local indicator = parent:createImage({
 		id = UIID.indicator[id],
-		path = INDICATOR_TEXTURE,
+		path = CONSTANTS.INDICATOR_TEXTURE,
 	})
-	indicator.color = INDICATOR_COLOR
+	indicator.color = CONSTANTS.INDICATOR_COLOR
 	indicator.absolutePosAlignX = absolutePosAlignX
 	indicator.absolutePosAlignY = absolutePosAlignY
 	indicator:setLuaData("indicatorID", id)
 	return indicator
 end
 
---- @param params openColorPickerMenu.new.params
+--- @param params tes3uiElement.createColorPicker.params
 --- @param picker ColorPicker
 --- @param parent tes3uiElement
 local function createPickerBlock(params, picker, parent)
@@ -325,8 +188,8 @@ local function createPickerBlock(params, picker, parent)
 		id = UIID.indicator.slider,
 		step = 1,
 		jump = 1,
-		current = mainIndicatorInitialAbsolutePosAlignX * SLIDER_SCALE,
-		max = SLIDER_SCALE,
+		current = mainIndicatorInitialAbsolutePosAlignX * CONSTANTS.SLIDER_SCALE,
+		max = CONSTANTS.SLIDER_SCALE,
 	})
 	slider.width = picker.mainWidth
 	slider.borderBottom = 8
@@ -397,20 +260,20 @@ local function createPickerBlock(params, picker, parent)
 		local x = math.clamp(e.relativeX, 1, mainPicker.width)
 		local y = math.clamp(e.relativeY, 1, mainPicker.height)
 		local pickedColor = picker.mainImage:getPixel(x, y)
-		colorSelected(picker, parent, pickedColor, picker.currentAlpha, currentPreview)
+		update.colorSelected(picker, parent, pickedColor, picker.currentAlpha, currentPreview)
 
 		x = x / mainPicker.width
 		y = y / mainPicker.height
 		mainIndicator.absolutePosAlignX = x
 		mainIndicator.absolutePosAlignY = y
-		slider.widget.current = x * SLIDER_SCALE
+		slider.widget.current = x * CONSTANTS.SLIDER_SCALE
 		mainRow:getTopLevelMenu():updateLayout()
 	end)
 	slider:register(tes3.uiEvent.partScrollBarChanged, function(e)
-		local x = math.clamp((slider.widget.current / SLIDER_SCALE) * mainPicker.width, 1, mainPicker.width)
+		local x = math.clamp((slider.widget.current / CONSTANTS.SLIDER_SCALE) * mainPicker.width, 1, mainPicker.width)
 		local y = mainIndicator.absolutePosAlignY * mainPicker.height
 		local pickedColor = picker.mainImage:getPixel(x, y)
-		colorSelected(picker, parent, pickedColor, picker.currentAlpha, currentPreview)
+		update.colorSelected(picker, parent, pickedColor, picker.currentAlpha, currentPreview)
 
 		x = x / mainPicker.width
 		y = y / mainPicker.height
@@ -432,7 +295,7 @@ local function createPickerBlock(params, picker, parent)
 		currentHSV.h = pickedHSV.h
 		pickedColor = oklab.hsvlib_hsv_to_srgb(currentHSV)
 
-		hueChanged(picker, parent, pickedColor, picker.currentAlpha, currentPreview, mainPicker)
+		update.hueChanged(picker, parent, pickedColor, picker.currentAlpha, currentPreview, mainPicker)
 
 		hueIndicator.absolutePosAlignY = y / huePicker.height
 		mainRow:getTopLevelMenu():updateLayout()
@@ -441,7 +304,7 @@ local function createPickerBlock(params, picker, parent)
 	if params.alpha then
 		alphaPicker:register(tes3.uiEvent.mouseStillPressed, function(e)
 			local y = math.clamp(e.relativeY / alphaPicker.height, 0, 1)
-			colorSelected(picker, parent, picker.currentColor, 1 - y, currentPreview)
+			update.colorSelected(picker, parent, picker.currentColor, 1 - y, currentPreview)
 			alphaIndicator.absolutePosAlignY = y
 			mainRow:getTopLevelMenu():updateLayout()
 		end)
@@ -450,11 +313,11 @@ local function createPickerBlock(params, picker, parent)
 	if params.showOriginal then
 		--- @param e tes3uiEventData
 		local function resetColor(e)
-			hueChanged(picker, parent, params.initialColor, params.initialAlpha, currentPreview, mainPicker)
+			update.hueChanged(picker, parent, params.initialColor, params.initialAlpha, currentPreview, mainPicker)
 
 			mainIndicator.absolutePosAlignX = mainIndicatorInitialAbsolutePosAlignX
 			mainIndicator.absolutePosAlignY = mainIndicatorInitialAbsolutePosAlignY
-			slider.widget.current = mainIndicatorInitialAbsolutePosAlignX * SLIDER_SCALE
+			slider.widget.current = mainIndicatorInitialAbsolutePosAlignX * CONSTANTS.SLIDER_SCALE
 			hueIndicator.absolutePosAlignY = hueIndicatorInitialAbsolutePosAlignY
 			if params.alpha then
 				alphaIndicator.absolutePosAlignY = 1 - params.initialAlpha
@@ -498,7 +361,7 @@ local function getInputValue(input)
 	return text
 end
 
---- @param params openColorPickerMenu.new.params
+--- @param params tes3uiElement.createColorPicker.params
 --- @param picker ColorPicker
 --- @param parent tes3uiElement
 --- @param onNewColorEntered fun(newColor: ffiImagePixel|ImagePixel, alpha: number)
@@ -529,7 +392,7 @@ local function createDataBlock(params, picker, parent, onNewColorEntered)
 	label.borderLeft = 4
 
 	local input = dataRow:createTextInput({
-		id = tes3ui.registerID("ColorPicker_data_row_value_input"),
+		id = UIID.textInput,
 		text = inputText,
 	})
 	input:setLuaData("hasAlpha", params.alpha or false)
@@ -540,7 +403,7 @@ local function createDataBlock(params, picker, parent, onNewColorEntered)
 	input:registerAfter(tes3.uiEvent.keyEnter, function(e)
 		local color = format.hexToPixel(getInputValue(input))
 		-- Update other parts of the Color Picker
-		updateValueInput(parent, color, color.a)
+		update.updateValueInput(parent, color, color.a)
 		onNewColorEntered(color, color.a)
 	end)
 
@@ -559,7 +422,7 @@ local function createDataBlock(params, picker, parent, onNewColorEntered)
 	}
 end
 
---- @param params openColorPickerMenu.new.params
+--- @param params tes3uiElement.createColorPicker.params
 --- @param parent tes3uiElement
 local function createColorPickerWidget(params, parent)
 	if (not params.alpha) or (not params.initialAlpha) then
@@ -569,11 +432,11 @@ local function createColorPickerWidget(params, parent)
 	premultiply.pixel(params.initialColor, params.initialAlpha)
 
 	local picker = ColorPicker:new({
-		mainWidth = PICKER_MAIN_WIDTH,
-		height = PICKER_HEIGHT,
-		hueWidth = PICKER_VERTICAL_COLUMN_WIDTH,
-		previewWidth = PICKER_PREVIEW_WIDTH,
-		previewHeight = PICKER_PREVIEW_HEIGHT,
+		mainWidth = CONSTANTS.PICKER_MAIN_WIDTH,
+		height = CONSTANTS.PICKER_HEIGHT,
+		hueWidth = CONSTANTS.PICKER_VERTICAL_COLUMN_WIDTH,
+		previewWidth = CONSTANTS.PICKER_PREVIEW_WIDTH,
+		previewHeight = CONSTANTS.PICKER_PREVIEW_HEIGHT,
 		initialColor = params.initialColor,
 		initialAlpha = params.initialAlpha,
 	})
@@ -593,9 +456,9 @@ local function createColorPickerWidget(params, parent)
 		--- @param newColor ffiImagePixel|ImagePixel
 		--- @param alpha number
 		local function onNewColorEntered(newColor, alpha)
-			hueChanged(picker, parent, newColor, alpha, pickers.currentPreview, pickers.mainPicker)
+			update.hueChanged(picker, parent, newColor, alpha, pickers.currentPreview, pickers.mainPicker)
 			-- TODO: looks like ffi can convert our ImagePixel table to ffiImagePixel when calling oklab.hsvlib_srgb_to_hsv
-			updateIndicatorPositions(parent, newColor, alpha)
+			update.updateIndicatorPositions(parent, newColor, alpha)
 		end
 
 		createDataBlock(params, picker, container, onNewColorEntered)
