+++ misc/package/Data Files/MWSE/core/mwse/ui/tes3uiElement/createColorPicker/uiid.lua
@@ -1,0 +1,11 @@
+local UIID = {
+	indicator = {
+		main = tes3ui.registerID("ColorPicker_main_picker_indicator"),
+		hue = tes3ui.registerID("ColorPicker_hue_picker_indicator"),
+		alpha = tes3ui.registerID("ColorPicker_alpha_picker_indicator"),
+		slider = tes3ui.registerID("ColorPicker_main_picker_slider"),
+	},
+	textInput = tes3ui.registerID("ColorPicker_data_row_value_input"),
+}
+
+return UIID
